Create separate packages for labs content
Instead of one large labs package. There are now individual packages for labs features. Moves `async-task` and `create-react-component` to their own packages.
