+++ lerna.json
@@ -1,4 +1,4 @@
 {
-  "packages": ["packages/*"],
+  "packages": ["packages/*", "packages/labs/*"],
   "version": "independent"
 }
