+++ packages/labs/async-task/.gitignore
@@ -1,5 +1,4 @@
 /development/
 /node_modules/
 /index.*
-/frameworks/
-/controllers/
+/async-task.*
