+++ packages/labs/async-task/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@lit/labs",
+  "name": "@lit-labs/async-task",
   "version": "1.0.0-pre.1",
-  "description": "An experimental playground for Lit helpers and tools. Your mileage may vary.",
+  "description": "A controller for Lit that renders asynchronous tasks.",
   "license": "BSD-3-Clause",
   "homepage": "https://lit-element.polymer-project.org/",
   "repository": "lit-html",
@@ -16,14 +16,6 @@
     ".": {
       "development": "./development/index.js",
       "default": "./index.js"
-    },
-    "./frameworks/": {
-      "development": "./development/frameworks/",
-      "default": "./frameworks/"
-    },
-    "./controllers/": {
-      "development": "./development/controllers/",
-      "default": "./controllers/"
     }
   },
   "files": [
@@ -33,23 +25,19 @@
     "/index.d.ts.map",
     "/index.js",
     "/index.js.map",
-    "/frameworks/react/create-component.d.ts",
-    "/frameworks/react/create-component.d.ts.map",
-    "/frameworks/react/create-component.js",
-    "/frameworks/react/create-component.js.map",
-    "/controllers/async-task.d.ts",
-    "/controllers/async-task.d.ts.map",
-    "/controllers/async-task.js",
-    "/controllers/async-task.js.map"
+    "/async-task.d.ts",
+    "/async-task.d.ts.map",
+    "/async-task.js",
+    "/async-task.js.map"
   ],
   "scripts": {
     "build": "npm run clean && tsc && rollup -c",
-    "clean": "rm -rf index.{js,js.map,d.ts} development/ controllers/ frameworks/",
+    "clean": "rm -rf index.{js,js.map,d.ts} development/",
     "dev": "scripts/dev.sh",
     "build:ts": "tsc",
     "build:ts:watch": "tsc --watch",
     "test": "npm run test:dev && npm run test:prod",
-    "test:dev": "cd ../tests && npx wtr '../labs/development/**/*_test.js'",
+    "test:dev": "cd ../../tests && npx wtr '../labs/async-task/development/**/*_test.js'",
     "test:prod": "TEST_PROD_BUILD=true npm run test:dev",
     "test:watch": "npm run test:dev -- --watch",
     "checksize": "rollup -c --environment=CHECKSIZE",
@@ -62,21 +50,14 @@
     "@types/chai": "^4.0.1",
     "@types/trusted-types": "^1.0.1",
     "@web/test-runner-mocha": "^0.3.5",
-    "@webcomponents/shadycss": "^1.8.0",
     "chokidar-cli": "^2.1.0",
     "concurrently": "^5.3.0",
     "mocha": "^8.1.3",
     "rollup": "^2.28.2",
-    "typescript": "^4.1.3",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "@types/react": "^16.9.50",
-    "@types/react-dom": "^16.9.8"
+    "typescript": "^4.1.3"
   },
   "dependencies": {
-    "@lit/reactive-element": "^1.0.0-pre.1",
-    "lit-element": "^3.0.0-pre.2",
-    "lit-html": "^2.0.0-pre.4"
+    "@lit/reactive-element": "^1.0.0-pre.1"
   },
   "publishConfig": {
     "access": "public"
