+++ packages/labs/async-task/rollup.config.js
@@ -1,0 +1,21 @@
+/**
+ * @license
+ * Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {litProdConfig} from '../../../rollup-common.js';
+
+export default litProdConfig({
+  classPropertyPrefix: '⍶',
+  entryPoints: ['index', 'async-task'],
+  external: ['@lit/reactive-element'],
+});
