+++ packages/labs/create-react-component/.gitignore
@@ -1,0 +1,4 @@
+/development/
+/node_modules/
+/index.*
+/create-react-component.*
