+++ packages/labs/create-react-component/package.json
@@ -1,0 +1,69 @@
+{
+  "name": "@lit-labs/create-react-component",
+  "version": "1.0.0-pre.1",
+  "description": "A React component wrapper for web components.",
+  "license": "BSD-3-Clause",
+  "homepage": "https://lit-element.polymer-project.org/",
+  "repository": "lit-html",
+  "type": "module",
+  "main": "index.js",
+  "module": "index.js",
+  "typings": "index.d.ts",
+  "directories": {
+    "test": "test"
+  },
+  "exports": {
+    ".": {
+      "development": "./development/index.js",
+      "default": "./index.js"
+    }
+  },
+  "files": [
+    "/src/",
+    "!/src/test/",
+    "/index.d.ts",
+    "/index.d.ts.map",
+    "/index.js",
+    "/index.js.map",
+    "/create-react-component.d.ts",
+    "/create-react-component.d.ts.map",
+    "/create-react-component.js",
+    "/create-react-component.js.map"
+  ],
+  "scripts": {
+    "build": "npm run clean && tsc && rollup -c",
+    "clean": "rm -rf index.{js,js.map,d.ts} development/",
+    "dev": "scripts/dev.sh",
+    "build:ts": "tsc",
+    "build:ts:watch": "tsc --watch",
+    "test": "npm run test:dev && npm run test:prod",
+    "test:dev": "cd ../../tests && npx wtr '../labs/create-react-component/development/**/*_test.js'",
+    "test:prod": "TEST_PROD_BUILD=true npm run test:dev",
+    "test:watch": "npm run test:dev -- --watch",
+    "checksize": "rollup -c --environment=CHECKSIZE",
+    "regen-package-lock": "rm -rf node_modules package-lock.json; npm install"
+  },
+  "author": "The Polymer Authors",
+  "devDependencies": {
+    "@esm-bundle/chai": "^4.1.5",
+    "@types/mocha": "^8.0.3",
+    "@types/chai": "^4.0.1",
+    "@types/trusted-types": "^1.0.1",
+    "@web/test-runner-mocha": "^0.3.5",
+    "chokidar-cli": "^2.1.0",
+    "concurrently": "^5.3.0",
+    "mocha": "^8.1.3",
+    "rollup": "^2.28.2",
+    "typescript": "^4.1.3",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "@types/react": "^16.9.50",
+    "@types/react-dom": "^16.9.8"
+  },
+  "dependencies": {
+    "@lit/reactive-element": "^1.0.0-pre.1"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
